display_and_save_categories

+++ admin-app/src/actions/auth.actions.js
@@ -12,11 +12,11 @@ export const login = (user) => {
     if (res.status === 200) {
       const { token, user } = res.data;
       localStorage.setItem("token", token);
-      localStorage.setItem('user',JSON.stringify(user))
+      localStorage.setItem("user", JSON.stringify(user));
       dispatch({
-        type:authConstants.LOGIN_SUCCESS,
-        payload:{token,user}
-      })
+        type: authConstants.LOGIN_SUCCESS,
+        payload: { token, user },
+      });
     } else {
       if (res.status === 400) {
         dispatch({
@@ -33,7 +33,7 @@ export const IsUserLoggedIn = () => {
   return async (dispatch) => {
     const token = localStorage.getItem("token");
     if (token) {
-      const user = JSON.parse(localStorage.getItem("user")) ;
+      const user = JSON.parse(localStorage.getItem("user"));
       dispatch({
         type: authConstants.LOGIN_SUCCESS,
         payload: {
@@ -41,23 +41,32 @@ export const IsUserLoggedIn = () => {
           user,
         },
       });
-    }else{
+    } else {
       dispatch({
         type: authConstants.LOGIN_FAILURE,
         payload: {
-          error: 'Failed to login',
+          error: "Failed to login",
         },
       });
     }
   };
 };
 
-
 export const signout = () => {
   return async (dispatch) => {
-    localStorage.clear();
-    dispatch({
-      type: authConstants.LOGOUT_REQUEST,
-    });
+    dispatch({ type: authConstants.LOGOUT_REQUEST });
+    const res = await axios.post("/admin/signout");
+
+    if (res.status === 200) {
+      localStorage.clear();
+      dispatch({
+        type: authConstants.LOGOUT_SUCCESS,
+      });
+    } else {
+      dispatch({
+        type: authConstants.LOGOUT_FAILURE,
+        payload:{error:res.data.error}
+      });
+    }
   };
 };
